feature: Refactor static catalog to use Compat pattern for cross-platform support
Refactored the static catalog implementation to use the Compat pattern for platform-specific
file I/O operations. This ensures compatibility across JVM, Scala.js, and Scala Native platforms.

Changes:
- Introduced FileIOCompat and StaticCatalogCompat traits for platform abstraction
- Moved file I/O dependent code to platform-specific directories (.jvm, .js, .native)
- Created stub implementations for Scala.js which has no file I/O support
- Updated Context to use Compat.loadStaticCatalog for catalog initialization
- Updated CLAUDE.md with platform-specific development guidelines

This allows the static catalog feature to compile and run on all platforms while
gracefully degrading to InMemoryCatalog on platforms without file I/O support.
